camelCasing

+++ src/components/SingleStreamManage.jsx
@@ -6,11 +6,16 @@ export default class SingleStreamManage extends Component {
   constructor(props){
     super(props);
     this.state={
-      stream:this.props.stream
+      stream:this.props.stream,
+      readyDelete: false
     }
   }
 
   deleteStream = async () => {
+    if(!this.state.readyDelete){
+      this.setState({readyDelete:true})
+      return
+    }
     const streamId = this.state.stream.id;
     try {
         await streams.delete_stream(streamId)
@@ -40,6 +45,14 @@ export default class SingleStreamManage extends Component {
     }
   }
 
+  onStopStart = async () =>{
+    if(this.state.stream.status==="active"){
+      await this.stopStream()
+      return
+    } 
+    await this.startStream()
+  }
+
   showRaw = () => {
     this.props.display(this.state.stream)
   }
@@ -50,7 +63,13 @@ export default class SingleStreamManage extends Component {
         <p>
           <b>{this.state.stream.streamName}</b>
           Status: {this.state.stream.status}
-          <button onClick={this.showRaw}>Show Raw</button>
+          <button className='single-button' onClick={this.showRaw}>Inspect</button>
+          <div style={{width:'100%', height:'0.1px', backgroundColor:'black'}}/>
+          <div style={{display:'flex', flexDirection:'row', justifyContent:'center'}}>
+            <button className='single-button' style={{color:'red'}} onClick={this.deleteStream}>{this.state.readyDelete?"Confirm":"Delete"}</button>
+            <button className='single-button' onClick={this.onStopStart}>{this.state.stream.status==="active"?"Stop":"Start"}</button>
+          </div>
+
         </p>
       </div>
     )
